services_1/real_time_analytics/requirements.txt
kafka-python==2.0.2
numpy==1.21.1
pandas==1.3.1